kas-k8s*: add seccomp DISTRO_FEATURE, cleanup
runc needs to be built with seccomp if the kernel supports it,
the PACKAGECONFIG is set if seccomp is in DISTRO_FEATURES.

Move common k8s settings into their own local.conf group, so
only the virtual/kernal PROVIDER need be overridden in rpi4 config.

hugetlb support is added in linux-yocto_virtualization.inc branch
with the k8s DISTRO_FEATURE set.

Signed-off-by: Tim Orling <ticotimo@gmail.com>